chore(plugin-server): Remove functional testss
